+++ plugins/common/parallel/parallel_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/metric"
-	"github.com/influxdata/telegraf/plugins/processors/reverse_dns/parallel"
+	"github.com/influxdata/telegraf/plugins/common/parallel"
 	"github.com/influxdata/telegraf/testutil"
 	"github.com/stretchr/testify/require"
 )
